fix: correctly detect PR number when run with workflow_run
Update the call to the GitHub Rest API, as the previous one wasn't able to retrieve the Pull Request
associated to the workflow_run.

From time to time, the API call may get rate limit errors given search API seems to use many calls
internally. In this case, the error is caught and a warning is logged.
Running the workflow again should then succeed.

Co-authored-by: benjaminparisel <benjamin.parisel@bonitasoft.com>
Co-authored-by: Thomas Bouffard <27200110+tbouffard@users.noreply.github.com>